Rename `cupy.manipulation` submodule to `cupy._manipulation`
