+++ cupy/_manipulation/dims.py
