+++ cupy/_manipulation/kind.py
