+++ cupy/_manipulation/rearrange.py
