+++ cupy/_manipulation/shape.py
