+++ cupy/_manipulation/split.py
