+++ cupy/_manipulation/transpose.py
