+++ cupy/_misc/memory_ranges.py
@@ -1,6 +1,6 @@
 from cupy.core import _kernel
 from cupy.core import _memory_range
-from cupy.manipulation import join
+from cupy._manipulation import join
 from cupy._sorting import search
