--- cupy/manipulation/__init__.py
@@ -1,14 +1,0 @@
-# Functions from the following NumPy document
-# https://docs.scipy.org/doc/numpy/reference/routines.array-manipulation.html
-
-# "NOQA" to suppress flake8 warning
-from cupy.manipulation import add_remove  # NOQA
-from cupy.manipulation import basic  # NOQA
-from cupy.manipulation import dims  # NOQA
-from cupy.manipulation import join  # NOQA
-from cupy.manipulation import kind  # NOQA
-from cupy.manipulation import rearrange  # NOQA
-from cupy.manipulation import shape  # NOQA
-from cupy.manipulation import split  # NOQA
-from cupy.manipulation import tiling  # NOQA
-from cupy.manipulation import transpose  # NOQA
